+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- `"SCMP_ARCH_MIPS64N32"` to `ScmpArch::from_str()`.
 
 ### Changed