+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+
+### Changed
+
+### Removed
+
+## 0.2.3 - 2022-02-14
+### Added
 - `"SCMP_ARCH_MIPS64N32"` to `ScmpArch::from_str()`.
 - `ScmpFilterContext::{get,set}_act_badarch()` to get/set the default action taken on a syscall for
 an architecture not in the filter.